Begin splitting ViewThemeUtils. Part 1: AndroidViewThemeUtils, for platform views
