+++ app/src/main/java/com/nextcloud/talk/activities/BaseActivity.kt
@@ -132,7 +132,7 @@ open class BaseActivity : AppCompatActivity() {
 
             val dialog = dialogBuilder.show()
 
-            viewThemeUtils.colorTextButtons(
+            viewThemeUtils.androidViewThemeUtils.colorTextButtons(
                 dialog.getButton(AlertDialog.BUTTON_POSITIVE),
                 dialog.getButton(AlertDialog.BUTTON_NEGATIVE)
             )
