+++ app/src/main/java/com/nextcloud/talk/activities/FullScreenTextViewerActivity.kt
@@ -104,7 +104,7 @@ class FullScreenTextViewerActivity : AppCompatActivity() {
         supportActionBar?.title = fileName
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
-        viewThemeUtils.themeStatusBar(this, binding.textviewToolbar)
+        viewThemeUtils.androidViewThemeUtils.themeStatusBar(this, binding.textviewToolbar)
 
         if (resources != null) {
             DisplayUtils.applyColorToNavigationBar(
