+++ app/src/main/java/com/nextcloud/talk/adapters/items/ContactItem.java
@@ -113,7 +113,7 @@ public class ContactItem extends AbstractFlexibleItem<ContactItem.ContactItemVie
         holder.binding.avatarDraweeView.setController(null);
 
         if (participant.getSelected()) {
-            viewThemeUtils.colorImageView(holder.binding.checkedImageView);
+            viewThemeUtils.androidViewThemeUtils.colorImageView(holder.binding.checkedImageView);
             holder.binding.checkedImageView.setVisibility(View.VISIBLE);
         } else {
             holder.binding.checkedImageView.setVisibility(View.GONE);
