+++ app/src/main/java/com/nextcloud/talk/adapters/items/GenericTextHeaderItem.java
@@ -74,7 +74,7 @@ public class GenericTextHeaderItem extends AbstractHeaderItem<GenericTextHeaderI
             Log.d(TAG, "We have payloads, so ignoring!");
         } else {
             holder.binding.titleTextView.setText(title);
-            viewThemeUtils.colorPrimaryTextViewElement(holder.binding.titleTextView);
+            viewThemeUtils.androidViewThemeUtils.colorPrimaryTextViewElement(holder.binding.titleTextView);
         }
     }
