+++ app/src/main/java/com/nextcloud/talk/polls/ui/PollLoadingFragment.kt
@@ -55,7 +55,7 @@ class PollLoadingFragment : Fragment() {
     ): View {
         binding = DialogPollLoadingBinding.inflate(inflater, container, false)
         binding.root.layoutParams.height = fragmentHeight
-        viewThemeUtils.colorCircularProgressBar(binding.pollLoadingProgressbar)
+        viewThemeUtils.androidViewThemeUtils.colorCircularProgressBar(binding.pollLoadingProgressbar)
         return binding.root
     }
