+++ app/src/main/java/com/nextcloud/talk/ui/dialog/ContactsBottomDialog.kt
@@ -59,7 +59,7 @@ class ContactsBottomDialog(
         setContentView(binding.root)
         window?.setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT)
 
-        viewThemeUtils.themeDialog(binding.root)
+        viewThemeUtils.androidViewThemeUtils.themeDialog(binding.root)
         executeEntryMenuController(bundle)
     }
